+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -182,7 +182,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KV355YM7S8YEHWDZNTRV1XQX"),
+  _id: z.string().default("01KV4V0CK2J8DNY18WVH46SN1Z"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: ListDatasourcesStatus$inboundSchema,