+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-14T13:29:29.915Z",
+    "2026-06-15T05:10:09.577Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {