+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3861,7 +3861,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KV355YA7G0PVTA4CFQG5C9PY"),
+  id: z.string().default("01KV4V0C7V5PE531ZMBXGD2M0V"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),